FIX : 문제 삭제 api 수정
1. 기존에 작성되어있던 코드는 문제 삭제 시 qstCounter의 카운팅값을 1
감소하도록 되어있었으나,
해당 문제가 마지막으로 counting된 문제가 아닐 경우 다음 문제 생성 시
중복값 발생 가능성을 확인하여 해당 기능 삭제하였습니다.
(문제를 삭제하여도 카운팅이 줄지 않습니다)

2. 삭제 실패 여부를 좀 더 명확하게 알 수 있도록 에러메세지를
   추가하였습니다.
문제 자체가 존재하지 않을 경우 404, 문제를 등록한 유저가 아닐 경우
403을 반환하도록 하였습니다.

+++ backend/routes/index.js
@@ -418,20 +418,29 @@ router.get("/question-detail", async (req, res) => {
 
 // 문제 삭제 
 router.delete("/question/delete", async (req,res) => {
-  const { _id, user_id, qualification_type } = req.query;
+  const { _id, user_id } = req.query;
   console.log(`delete question by _id: ${_id} and user_id: ${user_id}`);
   try{
+    // 문제 id로 문제 find
+    const question = await Question.qstModel.findOne({
+      _id: _id
+    });
+
+    // 문제 자체가 존재하지 않으면 404
+    if (!question) {
+      return res.status(404).json({ message: "Question not found" });
+    }
+    
+    if (question.user_id !== user_id) {
+      return res.status(403).json({ message: "user_id does not match" });
+    }
+
     const result = await Question.qstModel.deleteOne({
       _id: _id,
       user_id: user_id
     });
-
-    if (result.deletedCount === 1) {
-      // qstCounter 감소 
-      const findCount = await Question.qstCounter.findOne({ type: qualification_type });
-      findCount.count -= 1;
-      await findCount.save();
     
+    if (result.deletedCount === 1) {
       res.status(200).json({ message: "Question deleted successfully" });
     } else {
       res.status(404).json({ message: "Question not found" });
